[Refactor/#142] MyTitleScreen, MyTitleViewModel를 screens의 title 패키지로 이동
